feat(ChanPubSub): add single-channel pubsub implementation
The third implementation of a fan-out publish-subscribe pattern provided by
this module, ChanPubSub, is the most refined, and best handles very large
numbers of subscribers. It also has the best throughput, and, in many cases,
should be the preferred choice for a pubsub implementation. This implementation
is built on the existing ChanCaster implementation, makes use of newer Go
features including generics and iterators, and is lower-level than previous
implementations.

See also NewChanPubSub, ChanPubSub.SubscribeContext, and ChanPubSub.Send.